아티스트 그룹 등록 수정 경로 변경
