+++ Fiestar/src/main/resources/templates/common/adminArtist.html
@@ -66,10 +66,10 @@
               아티스트 등록
               <ul>
                 <li>
-                  <a th:href="@{/admin/regi}" id="a">아티스트 그룹 등록</a>
+                  <a th:href="@{/artistAdmin/regi}" id="a">아티스트 그룹 등록</a>
                 </li>
                 <li>
-                  <a th:href="@{/admin/artistGroupUpdate}" id="a">아티스트 그룹 수정</a>
+                  <a th:href="@{/artistAdmin/artistGroupUpdate}" id="a">아티스트 그룹 수정</a>
                 </li>
               </ul>
             </li>
